requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
torch>=1.9.0
easyocr
Pillow
Werkzeug>=2.0.0
Gunicorn
requests
flask-cors
pyopenssl